requirements.txt
tensorflow
numpy==1.24.3
apache-tvm
